# app/web/docs/openapi-generated.yaml
basePath: /
definitions:
  app_api_v1_admin.AdminLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  app_api_v1_admin.AdminLoginResponse:
    properties:
      account_type:
        type: string
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/go-template_domain_entities.User'
    type: object
...
  app_api_v1_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  go-template_domain_auth.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/go-template_domain_entities.User'
    type: object
  go-template_domain_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  go-template_domain_entities.AccountType:
    enum:
    - user
    - admin
    - super_admin
    type: string
    x-enum-varnames:
    - AccountTypeUser
    - AccountTypeAdmin
    - AccountTypeSuperAdmin
...
  go-template_domain_entities.User:
    properties:
      account_type:
        $ref: '#/definitions/go-template_domain_entities.AccountType'
      auth_provider:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/guilhermebr/go-template
  description: A Go template API built with Domain-Driven Design principles
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Template API
  version: 1.0.0
paths:
...
  /admin/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin user with privilege validation
      parameters:
      - description: Admin login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app_api_v1_admin.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_api_v1_admin.AdminLoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Admin login
      tags:
      - admin
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go-template_domain_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-template_domain_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app_api_v1_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/go-template_domain_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"